Dtos added

+++ src/main/java/zdz/joanna/travelagency/service/ClientService.java
@@ -36,16 +36,20 @@ public class ClientService implements ClientServiceInterface {
 
 	public List<ClientDto> getAllClients() {
 		List<Client> clients = clientRepository.findAll();
-		System.out.println("clients : " + clients);
 		return clients.stream().map(c -> mapper.map(c, ClientDto.class)).collect(Collectors.toList());
 	}
 
-	public void deleteClient(ClientDto clientDto) {
-		System.out.println("clientDto to be romoved : " + clientDto);
-		Client cl = clientRepository.findById(clientDto.getId())
-				.orElseThrow(() -> new IllegalArgumentException("Invalid client Id:" + clientDto.getId()));
-		System.out.println("client to be romoved : " + cl);
-		clientRepository.delete(cl);
+	public boolean deleteClient(ClientDto clientDto) {
+		if (clientDto != null) {
+			try {
+				Client cl = clientRepository.findById(clientDto.getId()).get();
+				clientRepository.delete(cl);
+			} catch (Exception e) {
+				return false;
+			}
+			return true;
+		}
+		return false;
 	}
 
 	public boolean isValid(ClientDto clientDto) {
